Project restruturing + linting
