--- src/js/launchers/LinuxGames.js
@@ -1,11 +1,0 @@
-async function findGames() {
-    var {
-        exec
-    } = require('child_process');
-
-    console.log(exec('which steam'))
-}
-
-module.exports = {
-	findGames,
-};
