+++ src/renderer/css/login.css
@@ -1,30 +1,29 @@
 @font-face {
     font-family: Comfortaa;
-    src: url('../../fonts/Comfortaa-SemiBold.ttf') format('truetype');
+    src: url('../../assets/fonts/Comfortaa-SemiBold.ttf') format('truetype');
 }
 
 @keyframes fadeInUp {
     from {
-        transform: translate3d(0, -20px, 0)
+        transform: translate3d(0, -20px, 0);
     }
 
     to {
         transform: translate3d(0, 0, 0);
-        opacity: 1
+        opacity: 1;
     }
 }
 
 @keyframes fadeInDown {
     from {
-        transform: translate3d(0, 10px, 0)
+        transform: translate3d(0, 10px, 0);
     }
 
     to {
         transform: translate3d(0, 0, 0);
-        opacity: 1
+        opacity: 1;
     }
 }
-
 
 * {
     box-sizing: border-box;
@@ -58,7 +57,6 @@ body {
 }
 
 .titlebar-icons {
-
     margin-right: 8px !important;
     margin-top: 8px !important;
 
@@ -72,12 +70,14 @@ body {
     border-radius: 50px/50px;
     border: solid 10px;
     -webkit-app-region: no-drag;
-    filter: invert(26%) sepia(0%) saturate(393%) hue-rotate(183deg) brightness(93%) contrast(90%);
+    filter: invert(26%) sepia(0%) saturate(393%) hue-rotate(183deg)
+        brightness(93%) contrast(90%);
     display: block;
 }
 
 .titlebar-exit:hover {
-    filter: invert(29%) sepia(93%) saturate(6374%) hue-rotate(353deg) brightness(92%) contrast(94%);
+    filter: invert(29%) sepia(93%) saturate(6374%) hue-rotate(353deg)
+        brightness(92%) contrast(94%);
 }
 
 .titlebar-max {
@@ -86,12 +86,14 @@ body {
     border-radius: 50px/50px;
     border: solid 10px;
     -webkit-app-region: no-drag;
-    filter: invert(26%) sepia(0%) saturate(393%) hue-rotate(183deg) brightness(93%) contrast(90%);
+    filter: invert(26%) sepia(0%) saturate(393%) hue-rotate(183deg)
+        brightness(93%) contrast(90%);
     display: block;
 }
 
 .titlebar-max:hover {
-    filter: invert(80%) sepia(51%) saturate(2452%) hue-rotate(348deg) brightness(102%) contrast(110%);
+    filter: invert(80%) sepia(51%) saturate(2452%) hue-rotate(348deg)
+        brightness(102%) contrast(110%);
 }
 
 .titlebar-min {
@@ -100,19 +102,21 @@ body {
     border-radius: 50px/50px;
     border: solid 10px;
     -webkit-app-region: no-drag;
-    filter: invert(26%) sepia(0%) saturate(393%) hue-rotate(183deg) brightness(93%) contrast(90%);
+    filter: invert(26%) sepia(0%) saturate(393%) hue-rotate(183deg)
+        brightness(93%) contrast(90%);
     display: block;
 }
 
 .titlebar-min:hover {
-    filter: invert(58%) sepia(76%) saturate(543%) hue-rotate(81deg) brightness(92%) contrast(97%);
+    filter: invert(58%) sepia(76%) saturate(543%) hue-rotate(81deg)
+        brightness(92%) contrast(97%);
 }
 
 h1 {
     font-weight: bold;
     margin: 0;
     margin-bottom: 20px;
-    margin-top: 10px
+    margin-top: 10px;
 }
 
 h2 {
@@ -146,7 +150,7 @@ button {
     border-radius: 10px;
     border: none;
     background-color: #9f40ff;
-    color: #FFFFFF;
+    color: #ffffff;
     font-size: 14px;
     vertical-align: middle;
     padding: 11px 45px;
@@ -165,7 +169,7 @@ button:hover {
 
 button.ghost {
     background-color: transparent;
-    border-color: #FFFFFF;
+    border-color: #ffffff;
     font-size: 14px;
 }
 
@@ -175,7 +179,7 @@ button.ghost2 {
     color: black;
     font-weight: bolder;
     background-color: transparent;
-    border-color: #FFFFFF;
+    border-color: #ffffff;
 }
 
 button.ignore {
@@ -191,7 +195,7 @@ button.ignore:hover {
 }
 
 .leftsection {
-    background-color: #FFFFFF;
+    background-color: #ffffff;
     display: flex;
     align-items: center;
     justify-content: center;
@@ -211,7 +215,6 @@ input {
     font-family: Comfortaa;
 }
 
-
 input:focus {
     outline: none;
 }
@@ -221,8 +224,7 @@ input:focus {
     -moz-border-radius: 20px;
     -webkit-border-radius: 20px;
     border-radius: 20px;
-    box-shadow: 0 14px 28px rgba(0, 0, 0, 0.25),
-        0 10px 10px rgba(0, 0, 0, 0.22);
+    box-shadow: 0 14px 28px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.22);
     position: relative;
     overflow: hidden;
     width: 950px;
@@ -262,7 +264,6 @@ input:focus {
 }
 
 @keyframes show {
-
     0%,
     49.99% {
         opacity: 0;
@@ -292,11 +293,11 @@ input:focus {
 }
 
 .overlay {
-    background: linear-gradient(to right, #6900D1, #912CEE);
+    background: linear-gradient(to right, #6900d1, #912cee);
     background-repeat: no-repeat;
     background-size: cover;
     background-position: 0 0;
-    color: #FFFFFF;
+    color: #ffffff;
     position: relative;
     left: -100%;
     height: 100%;
@@ -347,7 +348,7 @@ input:focus {
 
 .mx-1 {
     margin-right: 0.25rem !important;
-  margin-left: 0.25rem !important;
+    margin-left: 0.25rem !important;
 }
 
 .fadeInUp {
