+++ src/renderer/index.html
@@ -1,0 +1,192 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <link rel="stylesheet" type="text/css" href="css/style.css" />
+        <meta charset="UTF-8" />
+
+        <script async type="module" src="js/listeners.js"></script>
+        <script async type="text/javascript" src="js/ipc.js"></script>
+    </head>
+    <body>
+        <div class="titlebar">
+            <div class="titlebar-icons">
+                <div onclick="min_window()" class="titlebar-min mx-1"></div>
+                <div onclick="max_window()" class="titlebar-max mx-1"></div>
+                <div onclick="close_window()" class="titlebar-exit mx-1"></div>
+            </div>
+        </div>
+
+        <div id="loading-overlay">
+            <div class="spinner-content">
+                <img
+                    src="../assets/img/spinner.svg"
+                    alt="Loading..."
+                    class="loading-spinner"
+                />
+                <p class="fadeInDown">Loading your favorite games...</p>
+            </div>
+        </div>
+
+        <div class="bg">
+            <div class="homebox" id="home">
+                <div class="children fadeInUp">
+                    <img
+                        class="head-pic"
+                        src="https://www.powerpyx.com/wp-content/uploads/gta-3-definitive-edition-wallpaper.jpg"
+                    />
+                </div>
+                <div class="children fadeInDown">
+                    <div class="jump-back">
+                        <p>Most Played Games</p>
+                    </div>
+                </div>
+                <div class="children fadeInLeft">
+                    <div class="online-friends">
+                        <p>Online Friends</p>
+                    </div>
+                </div>
+            </div>
+
+            <div class="secondorybox" id="recent">
+                <p>Recently Played</p>
+            </div>
+
+            <div class="secondorybox" id="games">
+                <p>All Games</p>
+                <div id="gamesList" class="fadeInDown"></div>
+            </div>
+
+            <div class="secondorybox" id="favs">
+                <p>Favourites</p>
+            </div>
+
+            <div class="secondorybox" id="messages">
+                <p>Messages</p>
+            </div>
+
+            <div class="secondorybox" id="activity">
+                <p>Activity</p>
+            </div>
+
+            <div class="secondorybox" id="friends">
+                <p>Friends</p>
+            </div>
+
+            <div class="leftbar">
+                <div class="d-flex justify-content-center">
+                    <div>
+                        <div class="user-pfp">
+                            <label for="file"></label>
+                            <input
+                                id="file"
+                                type="file"
+                                accept="image/*"
+                                onchange="loadFile(event)"
+                            />
+                            <img
+                                src="https://cdn.discordapp.com/avatars/633730629560958976/5c1abedd641bb81ecc381696950a0b16.png?size=1024"
+                                alt="Avatar"
+                                width="80"
+                                height="80"
+                                id="output"
+                            />
+                        </div>
+                        <input
+                            class="username"
+                            id="text"
+                            type="text"
+                            value="DashCruft"
+                            spellcheck="false"
+                            maxlength="12"
+                        />
+                    </div>
+                </div>
+
+                <div class="side-tabs">
+                    <div class="d-flex justify-content-center pb-2">
+                        <div class="side-tab" id="home-btn">
+                            <div id="indicator"></div>
+                            <img
+                                src="../assets/img/home.svg"
+                                height="25"
+                                width="25"
+                            />
+                            <div class="side-tab-text">Home</div>
+                        </div>
+                    </div>
+
+                    <div class="category-name d-flex justify-content-left">
+                        Games
+                    </div>
+                    <div class="d-flex justify-content-center">
+                        <div class="side-tab" id="recent-btn">
+                            <img
+                                id="recently-btn-img"
+                                src="../assets/img/recent.svg"
+                                height="25"
+                                width="25"
+                            />
+                            <div class="side-tab-text">Recent</div>
+                        </div>
+                    </div>
+                    <div class="d-flex justify-content-center">
+                        <div class="side-tab" id="games-btn">
+                            <img
+                                id="games-btn-img"
+                                src="../assets/img/games.svg"
+                                height="25"
+                                width="25"
+                            />
+                            <div class="side-tab-text">All Games</div>
+                        </div>
+                    </div>
+                    <div class="d-flex justify-content-center pb-2">
+                        <div class="side-tab" id="favs-btn">
+                            <img
+                                id="favs-btn-img"
+                                src="../assets/img/favs.svg"
+                                height="25"
+                                width="25"
+                            />
+                            <div class="side-tab-text">Favourites</div>
+                        </div>
+                    </div>
+
+                    <div class="category-name d-flex justify-content-start">
+                        Friends
+                    </div>
+                    <div class="d-flex justify-content-center">
+                        <div class="side-tab" id="messages-btn">
+                            <img
+                                src="../assets/img/messages.svg"
+                                height="25"
+                                width="25"
+                            />
+                            <div class="side-tab-text">Messages</div>
+                        </div>
+                    </div>
+                    <div class="d-flex justify-content-center">
+                        <div class="side-tab" id="activity-btn">
+                            <img
+                                src="../assets/img/activity.svg"
+                                height="25"
+                                width="25"
+                            />
+                            <div class="side-tab-text">Activity</div>
+                        </div>
+                    </div>
+                    <div class="d-flex justify-content-center">
+                        <div class="side-tab" id="friends-btn">
+                            <img
+                                src="../assets/img/friends.svg"
+                                height="25"
+                                width="25"
+                            />
+                            <div class="side-tab-text">All Friends</div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </body>
+</html>
