+++ src/renderer/js/launchers/LinuxGames.js
@@ -1,0 +1,9 @@
+async function findGames() {
+    var { exec } = require('child_process');
+
+    console.log(exec('which steam'));
+}
+
+module.exports = {
+    findGames,
+};
