+++ src/renderer/login.html
@@ -1,0 +1,77 @@
+<!DOCTYPE html>
+<html lang="en">
+    <meta charset="UTF-8" />
+
+    <head>
+        <link rel="stylesheet" type="text/css" href="css/login.css" />
+
+        <script async type="module" src="js/login/listener.js"></script>
+        <script async type="text/javascript" src="js/login/ipc.js"></script>
+    </head>
+
+    <body>
+        <div class="titlebar">
+            <div class="titlebar-icons">
+                <div onclick="min_window()" class="titlebar-min mx-1"></div>
+                <div onclick="max_window()" class="titlebar-max mx-1"></div>
+                <div onclick="close_window()" class="titlebar-exit mx-1"></div>
+            </div>
+        </div>
+
+        <div class="bg">
+            <button onclick="load_main()" class="fadeInUp ignore">
+                Continue and Ignore
+            </button>
+            <div class="container fadeInUp" id="container">
+                <div class="form-container sign-up-container">
+                    <div class="leftsection">
+                        <h1>Create Account</h1>
+                        <input type="text" placeholder="Username" />
+                        <input type="email" placeholder="Email" />
+                        <input type="password" placeholder="Password" />
+                        <button onclick="load_main()">Sign Up</button>
+                    </div>
+                </div>
+                <div class="form-container sign-in-container">
+                    <div class="leftsection">
+                        <h1>Sign in</h1>
+                        <input type="text" placeholder="Username" />
+                        <input type="password" placeholder="Password" />
+                        <button onclick="load_main()">Sign In</button>
+                        <a>Forgot your password?</a>
+                    </div>
+                </div>
+                <div class="overlay-container">
+                    <div class="overlay">
+                        <div class="overlay-panel overlay-left">
+                            <h1 class="mt-4">Dear, User!</h1>
+                            <p>
+                                We believe and respect complete privacy. We do
+                                not sell your data! It is completely your choice
+                                to sign up or login. <br /><br />
+                                If you do not login or sign up, You will not
+                                have access to some integerated features in this
+                                launcher. <br /><br /><br />
+                                - Lazap Development Team
+                            </p>
+                            <button id="signIn">Sign In</button>
+                        </div>
+                        <div class="overlay-panel overlay-right">
+                            <h1 class="mt-4">Welcome Back!</h1>
+                            <p>
+                                We believe and respect complete privacy. We do
+                                not sell your data! It is completely your choice
+                                to sign up or login. <br /><br />
+                                If you do not login or sign up, You will not
+                                have access to some integerated features in this
+                                launcher. <br /><br /><br />
+                                - Lazap Development Team
+                            </p>
+                            <button id="signUp">Sign Up</button>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </body>
+</html>
